@@ -1,25 +1,31 @@
 {
   "name": "v-viewer",
+  "scope": "VueViewer",
   "version": "3.0.0",
   "description": "Image viewer component for vue, supports rotation, scale, zoom and so on, based on viewer.js",
-  "main": "dist/v-viewer.common.js",
-  "unpkg": "dist/v-viewer.umd.min.js",
-  "types": "types/index.d.ts",
+  "main": "./dist/index.umd.js",
+  "module": "./dist/index.es.js",
+  "types": "./types/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.es.js",
+      "require": "./dist/index.umd.js"
+    }
+  },
   "scripts": {
-    "serve-example": "vue-cli-service serve",
-    "build-example": "vue-cli-service build --mode example",
-    "build": "vue-cli-service build --mode release --target lib --name v-viewer ./src/main.ts",
-    "release": "standard-version",
-    "lint": "vue-cli-service lint",
-    "lint fix": "vue-cli-service lint --fix",
-    "ui": "vue ui",
-    "gh": "node deploy",
-    "publish": "npm publish"
+    "dev": "npm -C example run dev",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview:serve": "npm -C example run serve",
+    "preview:build": "npm -C example run build",
+    "commit": "npx git-cz",
+    "release-f": "standard-version -f",
+    "release-major": "standard-version -r major",
+    "release-minor": "standard-version -r minor",
+    "release-patch": "standard-version -r patch"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mirari/v-viewer.git"
-  },
+  "files": [
+    "dist"
+  ],
   "keywords": [
     "vue",
     "viewer",
@@ -28,52 +34,45 @@
     "img",
     "image"
   ],
-  "author": "mirari",
   "license": "MIT",
+  "author": "mirari",
   "bugs": {
     "url": "https://github.com/mirari/v-viewer/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mirari/v-viewer.git"
   },
   "homepage": "https://github.com/mirari/v-viewer#readme",
   "dependencies": {
     "lodash": "^4.17.21",
     "viewerjs": "^1.9.0",
-    "vue": "^3.0.0"
+    "vue": "^3.0.5"
   },
   "devDependencies": {
-    "@commitlint/cli": "^12.1.4",
-    "@commitlint/config-conventional": "^12.1.4",
-    "@types/lodash": "^4.14.168",
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-standard": "^5.1.2",
-    "@vue/eslint-config-typescript": "^7.0.0",
+    "@antfu/eslint-config": "^0.6.4",
+    "@commitlint/cli": "^12.1.1",
+    "@commitlint/config-conventional": "^12.1.1",
+    "@types/lodash-es": "^4.17.4",
+    "@typescript-eslint/eslint-plugin": "^4.22.1",
+    "@vitejs/plugin-vue": "^1.2.2",
+    "@vue/compiler-sfc": "^3.0.5",
     "bulma": "^0.9.2",
-    "core-js": "^3.6.5",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^7.0.0",
-    "lint-staged": "^11.0.0",
-    "sass": "^1.26.5",
-    "sass-loader": "^8.0.2",
+    "commitizen": "^4.2.3",
+    "create-banner": "^2.0.0",
+    "cz-conventional-changelog": "^3.3.0",
+    "eslint": "^7.25.0",
+    "husky": "^4.3.8",
+    "lint-staged": "^10.5.4",
+    "sass": "^1.32.13",
     "standard-version": "^9.3.0",
-    "typescript": "~4.1.5"
+    "typescript": "^4.1.3",
+    "vite": "^2.2.3",
+    "vue-tsc": "^0.0.24"
   },
-  "gitHooks": {
-    "pre-commit": "lint-staged",
-    "commit-msg": "commitlint -e $GIT_PARAMS"
-  },
-  "lint-staged": {
-    "*.{js,jsx,vue,ts,tsx}": [
-      "vue-cli-service lint",
-      "git add"
-    ]
+  "config": {
+    "commitizen": {
+      "path": "cz-conventional-changelog"
+    }
   }
 }
