@@ -1,31 +1,49 @@
-import Viewer from 'viewerjs'
+import ViewerJs from 'viewerjs'
 import debounce from 'lodash/debounce'
-import { App, nextTick } from 'vue'
+import { nextTick, watch } from 'vue'
+import type { Directive, DirectiveBinding, VNode } from 'vue'
 
-const install = (app: App, { name = 'viewer', debug = false }) => {
-  async function createViewer (el: HTMLElement, options: Viewer.Options, rebuild = false) {
-    // nextTick执行，否则可能漏掉未渲染完的子元素
+export type IcreateViewer = (el: HTMLElement, options: ViewerJs.Options, rebuild: boolean) => void
+
+const directive = ({ name = 'viewer', debug = false }) => {
+  async function createViewer(el: HTMLElement, options: ViewerJs.Options, rebuild = false) {
     await nextTick()
     if (rebuild || !el[`$${name}`]) {
       destroyViewer(el)
-      el[`$${name}`] = new Viewer(el, options)
+      el[`$${name}`] = new ViewerJs(el, options)
       log('viewer created')
-    } else {
+    }
+    else {
       el[`$${name}`].update()
       log('viewer updated')
     }
   }
 
-  function createObserver (el: HTMLElement, options: Viewer.Options, debouncedCreateViewer, rebuild: boolean) {
+  function createObserver(el: HTMLElement, options: ViewerJs.Options, debouncedCreateViewer: IcreateViewer, rebuild: boolean) {
     destroyObserver(el)
-    const MutationObserver = global.MutationObserver || global.WebKitMutationObserver || global.MozMutationObserver
+    const MutationObserver = window.MutationObserver || window.WebKitMutationObserver || window.MozMutationObserver
     if (!MutationObserver) {
       log('observer not supported')
       return
     }
-    const observer = new MutationObserver(function (mutations) {
-      mutations.forEach(function (mutation) {
-        log('viewer mutation:' + mutation.type)
+    const observer = new MutationObserver((mutations) => {
+      const matchImage = el.innerHTML.match(/<img([\w\W]+?)[\\/]?>/g)
+      // When there is no image, it is not recreated.
+      if (matchImage === null) {
+        log('observer no image')
+        return
+      }
+      // When there is no change, it is not recreated.
+      const $viewerNewImage = matchImage.join('')
+      if (el.$viewerOldImage === $viewerNewImage) {
+        log('observer no change')
+        return
+      }
+      else {
+        el.$viewerOldImage = $viewerNewImage
+      }
+      mutations.forEach((mutation) => {
+        log(`viewer mutation:${mutation.type}`)
         debouncedCreateViewer(el, options, rebuild)
       })
     })
@@ -35,22 +53,20 @@ const install = (app: App, { name = 'viewer', debug = false }) => {
     log('observer created')
   }
 
-  function createWatcher (el: HTMLElement, { expression }, vnode, debouncedCreateViewer) {
-    const simplePathRE = /^[A-Za-z_$][\w$]*(?:\.[A-Za-z_$][\w$]*|\['[^']*?']|\["[^"]*?"]|\[\d+]|\[[A-Za-z_$][\w$]*])*$/
-    if (!expression || !simplePathRE.test(expression)) {
-      log('only simple dot-delimited paths can create watcher')
-      return
-    }
-    el.$viewerUnwatch = vnode.context.$watch(expression, function (newVal, oldVal) {
-      log('change detected by watcher: ', expression)
+  function createWatcher(el: HTMLElement, binding: DirectiveBinding, vnode: VNode, debouncedCreateViewer: IcreateViewer) {
+    // const simplePathRE = /^[A-Za-z_$][\w$]*(?:\.[A-Za-z_$][\w$]*|\['[^']*?']|\["[^"]*?"]|\[\d+]|\[[A-Za-z_$][\w$]*])*$/
+    // if (!binding.value || !simplePathRE.test(binding.value)) {
+    //   log('only simple dot-delimited paths can create watcher')
+    //   return
+    // }
+    el.$viewerUnwatch = watch(() => binding.value, (newVal, oldVal) => {
+      log('change detected by watcher: ', binding.value)
       debouncedCreateViewer(el, newVal, true)
-    }, {
-      deep: true,
-    })
-    log('watcher created, expression: ', expression)
+    }, { deep: true })
+    log('watcher created, expression: ', binding.value)
   }
 
-  function destroyViewer (el: HTMLElement) {
+  function destroyViewer(el: HTMLElement) {
     if (!el[`$${name}`]) {
       return
     }
@@ -59,7 +75,7 @@ const install = (app: App, { name = 'viewer', debug = false }) => {
     log('viewer destroyed')
   }
 
-  function destroyObserver (el: HTMLElement) {
+  function destroyObserver(el: HTMLElement) {
     if (!el.$viewerMutationObserver) {
       return
     }
@@ -68,7 +84,7 @@ const install = (app: App, { name = 'viewer', debug = false }) => {
     log('observer destroyed')
   }
 
-  function destroyWatcher (el: HTMLElement) {
+  function destroyWatcher(el: HTMLElement) {
     if (!el.$viewerUnwatch) {
       return
     }
@@ -77,26 +93,24 @@ const install = (app: App, { name = 'viewer', debug = false }) => {
     log('watcher destroyed')
   }
 
-  function log (...args: any[]) {
+  function log(...args: any[]) {
     debug && console.log(...args)
   }
 
-  Vue.directive('viewer', {
-    bind (el: HTMLElement, binding, vnode) {
+  const directive: Directive = {
+    mounted(el, binding, vnode) {
       log('viewer bind')
       const debouncedCreateViewer = debounce(createViewer, 50)
       debouncedCreateViewer(el, binding.value)
-
       // 创建watch监听options表达式变化
       createWatcher(el, binding, vnode, debouncedCreateViewer)
-
       // 是否监听dom变化
       if (!binding.modifiers.static) {
         // 增加dom变化监听
         createObserver(el, binding.value, debouncedCreateViewer, binding.modifiers.rebuild)
       }
     },
-    unbind (el: HTMLElement, binding) {
+    unmounted(el) {
       log('viewer unbind')
       // 销毁dom变化监听
       destroyObserver(el)
@@ -105,9 +119,9 @@ const install = (app: App, { name = 'viewer', debug = false }) => {
       // 销毁viewer
       destroyViewer(el)
     },
-  })
+  }
+
+  return directive
 }
 
-export default {
-  install,
-}
+export default directive
