@@ -1,19 +1,19 @@
 import { App } from 'vue'
-import Viewer from 'viewerjs'
+import ViewerJs from 'viewerjs'
 
 declare namespace VueViewer {
   export interface InstallationOptions {
-    name: string;
-    debug: boolean;
-    defaultOptions: Viewer.Options;
+    name: string
+    debug: boolean
+    defaultOptions: ViewerJs.Options
   }
 
-  export function install(app: App, options?: InstallationOptions): void;
+  export function install(app: App, options?: InstallationOptions): void
 
-  export function setDefaults(defaultOptions: Viewer.Options): void;
+  export function setDefaults(defaultOptions: ViewerJs.Options): void
 
   export {
-    Viewer,
+    ViewerJs,
   }
 }
 
